Attempted to apply Spline Approximations to predict stock prices
It didn't work because I failed to realize that the final spline with all points accounted for is not the same as the spline that gets created as a function of time. As we get more data, the spline becomes more accurate between points. This means that if we build a spline as time goes on, the spline won't be able to accurately predict the next datapoint in time. Sad day for me